move all questions routes out of main
